Added /undo and cleaned up /error support
Both Twine and Ink laoders now support the /undo route. For Twine, functionality is limited to the Harlowe and SugarCube story formats. For Snowman story format in Twine as well as Ink overall, undo functionality is turned off and an error message is created when attempting to usee it. (This same message is returned when undoing is tried when no history exists in Harlowe and SugarCube story formats in Twine.)

Instead of sending messages to root route, a new /error one was created. Any errors that happen within the rendered wndow are relayed through main process and presented on the channel. The route will retain the last error to occur until the next 'click' action is taken. At which point, a new error will be shown or the empty string when no errors have occured.