ES6 and Error Dialogs
Moved more of the code over to ES6.

Error dialogs now happen based on main process events. If the settings.json file is missing, an error dialog occurs. The same with trying to test for various types of values in settings.json including if the port number is actually a number.