Added /redo
The /redo route acts like /undo and will "redo" the last action in the history stack if the story format provides that functionaity. (This is limited to Harlowe and SugarCube story formats.)

Ink as well as the Snowman story format in Twine do not provide redo functionality and thus any attempts to access it will cause an error to be generated.